requirements.txt
aniso8601==9.0.1
area==1.1.1
attrs==23.2.0
bitarray==2.9.3
bitstring==4.2.3
blinker==1.7.0
bottle==0.12.25
brother-ql==0.9.4
cachetools==5.3.3
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
dataclasses==0.6
db-dtypes==1.2.0
ecdsa==0.19.0
esptool==4.8.1
Flask==3.0.3
Flask-Cors==4.0.1
Flask-RESTful==0.3.10
fonttools==4.50.0
future==1.0.0
geographiclib==2.0
geojson==3.1.0
geopandas==1.0.1
geopy==2.4.1
google-api-core==2.17.1
google-auth==2.28.2
google-cloud-bigquery==3.19.0
google-cloud-core==2.4.1
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
grpcio==1.62.1
grpcio-status==1.62.1
idna==3.6
iniconfig==2.0.0
intelhex==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
MouseInfo==0.1.3
mpremote==1.23.0
numpy==1.26.4
packaging==24.0
packbits==0.6
pandas==2.2.1
Pillow==9.5.0
pluggy==1.5.0
protobuf==4.25.3
pyarrow==15.0.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.22
PyGetWindow==0.0.9
Pygments==2.17.2
PyMsgBox==1.0.9
pyogrio==0.9.0
pyparsing==3.1.2
pyperclip==1.9.0
pyproj==3.6.1
PyRect==0.2.0
PyScreeze==1.0.1
pyserial==3.5
pytest==8.3.2
python-dateutil==2.9.0.post0
pytweening==1.2.0
pytz==2024.1
pyusb==1.2.1
PyYAML==6.0.2
reedsolo==1.7.0
requests==2.31.0
rsa==4.9
scipy==1.13.0
setuptools==69.1.1
shapely==2.0.3
six==1.16.0
turfpy==0.0.7
tzdata==2024.1
urllib3==2.2.1
utm==0.7.0
vedo==2024.5.1
vtk==9.3.0
welltrajconvert==0.0.2
Werkzeug==3.0.2
wheel==0.43.0
